Adds comments
