+++ app/dashboard/routes/template-editor/save/previewPath.js
@@ -1,0 +1,14 @@
+const Template = require("template");
+
+module.exports = function (req, res, next) {
+  if (!req.body || !req.body.previewPath) return next();
+  Template.update(
+    req.blog.id,
+    req.params.templateSlug,
+    { previewPath: req.body.previewPath },
+    function (err) {
+      if (err) return next(err);
+      res.message(req.baseUrl + req.url, "Success!");
+    }
+  );
+};
