+++ app/dashboard/views/template-editor/preview.html
@@ -3,42 +3,28 @@
 	src="{{{preview}}}"
 ></iframe>
 <script type="text/javascript">
-function init_content_monitor() {
-    var content = document.querySelector('iframe');
+var content = document.querySelector("iframe");
 
-    // Listen to messages sent from the content iframe
-    var receiveMessage = function receiveMessage(e){
-      
-      console.log('here', e.origin, "{{{previewOrigin}}}")
-        // // Only react to messages from same domain as current document
-        if (e.origin !== "{{{previewOrigin}}}") return;
+// Listen to messages sent from the iframe which contains
+// the preview of the template. We inject the script
+// which sends these messages before the </body> tag of 
+// all HTML pages rendered on preview subdomains.
+var receiveMessage = function receiveMessage(e) {
+    // Only react to messages from the preview subdomain
+    // The user can click on links and load external pages
+    if (e.origin !== "{{{previewOrigin}}}") return;
 
-        // Handle the message
-        let path = e.data.slice('iframe:'.length);
+    // Extract the path of the page they are viewing
+    let path = e.data.slice("iframe:".length);
 
-var http = new XMLHttpRequest();
-var url = "{{{base}}}/previewPath";
-var params = 'previewPath=' + path;
-http.open('POST', url, true);
+    // Save the path they have viewed on the server
+    var http = new XMLHttpRequest();
+    var url = "{{{base}}}/settings";
+    var params = "previewPath=" + path;
+    http.open("POST", url, true);
+    http.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
+    http.send(params);
+};
 
-//Send the proper header information along with the request
-http.setRequestHeader('Content-type', 'application/x-www-form-urlencoded');
-
-http.onreadystatechange = function() {//Call a function when the state changes.
-    if(http.readyState == 4 && http.status == 200) {
-
-        // http.responseText will be anything that the server return
-        console.log(http.responseText);
-    }
-}
-http.send(params);
-
-        console.log(path);
-            };
-    window.addEventListener("message", receiveMessage, false);
-
-  }  
-
-  init_content_monitor();
-
+window.addEventListener("message", receiveMessage, false);
 </script>
